user profile converted to mvvm
