+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,9 @@
         android:theme="@style/Theme.MyApplication"
         tools:targetApi="31">
         <activity
+            android:name=".ui.profile.ProfileActivity"
+            android:exported="false" />
+        <activity
             android:name=".ui.profile.OtherUsersProfileActivity"
             android:exported="false" />
         <activity
